+++ docs/changes.md
@@ -1,4 +1,6 @@
 # 最新动态
+2022/01/06
+  * fscript中增加is\_int/is\_odd/is\_even等函数
 
 2022/01/05
   * fscript公开函数value\_id/value\_func。